requirements.txt

streamlit>=1.36
numpy>=1.24
pandas>=2.0
Pillow>=10.0
opencv-python-headless>=4.8
